调整策略编号的数据类型为uint64, 修改权限注册接口入参为uint64, 和权限模块保持一致

+++ config/config_trader.go
@@ -95,7 +95,7 @@ func TraderConfig() TraderParameter {
 }
 
 // GetStrategyParameterByCode 通过策略编码查找规则
-func GetStrategyParameterByCode(strategyCode int) *StrategyParameter {
+func GetStrategyParameterByCode(strategyCode uint64) *StrategyParameter {
 	strategies := TraderConfig().Strategies
 	for _, v := range strategies {
 		if v.Auto && v.Id == strategyCode {
