+++ storages/stockpool.go
@@ -14,7 +14,7 @@ type StockPool struct {
 	Rate           float64        `name:"涨跌幅%" dataframe:"rate"`
 	Buy            float64        `name:"委托价格" dataframe:"buy"`
 	Sell           float64        `name:"目标价格" dataframe:"sell"`
-	StrategyCode   int            `name:"策略编码" dataframe:"strategy_code"`
+	StrategyCode   uint64         `name:"策略编码" dataframe:"strategy_code"`
 	StrategyName   string         `name:"策略名称" dataframe:"strategy_name"`
 	Rules          uint64         `name:"规则" dataframe:"rules"`
 	BlockType      string         `name:"板块类型" dataframe:"block_type"`
